refactoring and fixing minor bugs
