+++ src/App.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Route, Routes, useNavigate } from "react-router-dom";
 import { auth, db } from "./lib/firebase";
 
-import AnimatedButton from "./components/AnimatedButton";
+import AnimatedButton from "./reusableComponents/AnimatedButton";
 import { FaArrowCircleUp } from "react-icons/fa";
 import Favorite from "./components/Favorite";
 import Loader from "./components/Loader";
@@ -13,7 +13,7 @@ import NotFoundPage from "./components/NotFound";
 import Post from "./components/Post";
 import PostView from "./pages/PostView";
 import Profile from "./pages/Profile";
-import ShareModal from "./components/ShareModal";
+import ShareModal from "./reusableComponents/ShareModal";
 import SideBar from "./components/SideBar";
 import SignupScreen from "./pages/Signup";
 import { makeStyles } from "@mui/styles";
@@ -265,7 +265,7 @@ function App() {
                   alignItems: "center",
                 }}
               >
-                <SideBar user={user} />
+                <SideBar />
                 <div
                   style={
                     !loadingPosts
